[style] #90 State 분리
